In __init__() of the 'TAF' class, the list _weather_groups has been overwritten by the dictionary returned by the function _parse_metar(). In order to keep the processing of METAR and TAF as compatible as possible, the dictionary returned by _parse_metar() is now appended to the _weather_group list. A METAR string only contains one "weather group" so the list seems kinda senseless here, but treating TAFs and METARs equally makes the programmer's life a bit easier.
decode_taf() is now a bit more compact. Every check for a certain group depends now on the type of input to parse. Each checked group can only be existent in a TAF, a METAR or both.

I extended the METAR specific decoding function _decode_temperature() as well. A parameter for selecting the unit has been added. 'unit' can be either 'C' or 'F'. Default is 'C'.